added shopping api
