+++ package.json
@@ -21,10 +21,12 @@
     "type": "git",
     "url": "git@github.com:ajay2507/ebay-node-api.git"
   },
+  "dependencies": {
+    "make-string": "^1.0.2"
+  },
   "devDependencies": {
     "chai": "^4.1.2",
     "eslint": "^5.8.0",
-    "make-string": "^1.0.2",
     "mocha": "^5.0.1",
     "nock": "^9.2.3",
     "sinon": "^4.4.5"
