+++ src/index.js
@@ -60,36 +60,11 @@ Ebay.prototype = {
         )
     },
 
-    getAllCategories: function (categoryID) {
-        this.options.name = categoryID ? categoryID : -1;
-        this.options.operationName = "GetCategoryInfo";
-        this.options.param = "CategoryID";
-        const url = urlObject.buildShoppingUrl(this.options);
-        return getRequest(url).then((data) => {
-            return JSON.parse(data);
-        }, console.error
-        )
-    },
-
     getVersion: function () {
         this.options.operationName = "getVersion";
         const url = urlObject.buildSearchUrl(this.options);
-        console.log(url);
         return getRequest(url).then((data) => {
             return JSON.parse(data)["getVersionResponse"][0];
-        }, console.error
-        )
-    },
-
-    getUserDetails: function (userID) {
-        if (!userID) throw new Error("User ID is null or invalid");
-        this.options.operationName = "GetUserProfile";
-        this.options.param = "UserID";
-        this.options.name = userID;
-        this.options.includeSelector = this.options.details ? "Details" : null;
-        const url = urlObject.buildShoppingUrl(this.options);
-        return getRequest(url).then((data) => {
-            return JSON.parse(data);
         }, console.error
         )
     },
